Give users more control and fix our HTTPX clients handling
Users can now use Notion clients in two ways: either as context managers, that
will always ensure a valid HTTPX client and will close it automatically, or
manually, which allows for a custom HTTPX client to be passed in but requires
that it's closed explicitely when using AsynClient.

This fixes #73 overall by not closing AsyncClient.client after the first
request and transfering the responsability of context management to the user.